added getting messages for a group #18

+++ SECP-service/src/main/java/com/visucius/secp/models/Group.java
@@ -58,7 +58,7 @@ public class Group {
 
     public long getId(){return id;}
 
-    public void setId(int id){this.id = id;}
+    public void setId(long id){this.id = id;}
 
     public void setName(String name) {
         this.name = name;
